module4-task2: - карта на экране предложения - список карточек
