+++ project/src/components/main-screen/main-screen.jsx
@@ -5,11 +5,7 @@ import OffersList from '../offers-list/offers-list';
 import cardOfferProp from '../offer-card/offer-card.prop';
 import InteractiveMap from '../interactive-map/interactive-map';
 
-const city = {
-  latitude: 52.38333,
-  longitude: 4.9,
-  zoom: 10,
-};
+import city from '../../mocks/city';
 
 function MainScreen({offersCount, offers}) {
   return (
@@ -82,10 +78,28 @@ function MainScreen({offersCount, offers}) {
         </div>
         <div className="cities">
           <div className="cities__places-container container">
-            <OffersList
-              offersCount={offersCount}
-              offers={offers}
-            />
+            <section className="cities__places places">
+              <h2 className="visually-hidden">Places</h2>
+              <b className="places__found">{offersCount} places to stay in Amsterdam</b>
+              <form className="places__sorting" action="#" method="get">
+                <span className="places__sorting-caption">Sort by</span>
+                <span className="places__sorting-type" tabIndex="0"> Popular
+                  <svg className="places__sorting-arrow" width="7" height="4">
+                    <use xlinkHref="#icon-arrow-select"></use>
+                  </svg>
+                </span>
+                <ul className="places__options places__options--custom places__options--opened">
+                  <li className="places__option places__option--active" tabIndex="0">Popular</li>
+                  <li className="places__option" tabIndex="0">Price: low to high</li>
+                  <li className="places__option" tabIndex="0">Price: high to low</li>
+                  <li className="places__option" tabIndex="0">Top rated first</li>
+                </ul>
+              </form>
+              <OffersList
+                className="cities__places-list tabs__content"
+                offers={offers}
+              />
+            </section>
             <div className="cities__right-section">
               <section className="cities__map map">
                 <InteractiveMap
@@ -95,7 +109,6 @@ function MainScreen({offersCount, offers}) {
               </section>
             </div>
           </div>
-
         </div>
       </main>
     </div>
