+++ project/src/mocks/city.js
@@ -1,0 +1,7 @@
+const city = {
+  latitude: 52.38333,
+  longitude: 4.9,
+  zoom: 10,
+};
+
+export default city;
